rename (files): fix filename style
rename filenames styles of Button component
