+++ src/components/Button/Button.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import Props from './button.props';
+import Props from './Button.props';
 import '../Text/index'
-import {getClassName}   from "./button.classes";
+import {getClassName}   from "./Button.classes";
 
 const Button: React.FC<Props> = ({ label, disabled = false, variant = 'filled', className = '', ...props }) => {
     return (
